Build pitch array dynamically
This way the pitch error propagates better.

Though I chose 880 Hz as the base point (instead of 440 Hz because the
upstream base is one octave higher from 440), it should be mostly
compatible with the original composer.

Notably, writing everything in 0th octave actually sounds kinda off, as
alluded to in the previous commit; this sounds fine meanwhile 🤷